requirements.txt
aiohttp==3.7.4.post0
anytree==2.8.0
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.4.0
Beaker==1.11.0
beautifulsoup4==4.10.0
btrfsutil==5.16.2
CacheControl==0.12.6
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==36.0.2
discord.py==1.7.3
distlib==0.3.4
distro==1.7.0
docopt==0.6.2
filelock==3.4.2
fpdf==1.7.2
greenlet==1.1.2
html5lib==1.1
idna==3.3
importlib-metadata==4.8.1
keyutils==0.6
lensfun==0.3.3
lxml==4.8.0
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.1.1
meson==0.62.0
Mimeo==2021.11.20.2.28.56
more-itertools==8.10.0
msgpack==1.0.3
multidict==6.0.2
ordered-set==4.0.2
packaging==20.9
pep517==0.12.0
Pillow==9.0.1
pipenv==2022.3.24
platformdirs==2.5.0
ply==3.11
progress==1.6
pycparser==2.21
Pygments==2.11.2
PyGObject==3.42.0
PyNaCl==1.4.0
pynvim==0.4.3
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyxdg==0.27
PyYAML==6.0
ranger-fm==1.9.3
Reflector==2021.11.20.2.41.3
requests==2.27.1
resolvelib==0.5.5
retrying==1.3.3
SCons==4.3.0
six==1.16.0
soupsieve==2.3.1
TBB==0.2
team==1.0
toml==0.10.2
tomli==2.0.1
typing-extensions==4.1.1
udiskie==2.4.2
urllib3==1.26.9
virtualenv==20.11.0
virtualenv-clone==0.5.7
webencodings==0.5.1
wikipedia==1.4.0
yarl==1.7.2
zipp==3.7.0
